+++ nikamap/contmap.py
@@ -895,6 +895,7 @@ class ContMap(NDDataArray):
         psf=True,
         fixed_psf=True,
         background=True,
+        local_background=False,
         background_clipping=3,
         grouping_threshold=3,
     ):
@@ -911,7 +912,9 @@ class ContMap(NDDataArray):
         fixed_psf : bool, optional,
             Fix sources positions in the psf fit, default True
         background : bool, optional
-            Estimate and remove a background, by default True
+            Estimate and remove a global median background, by default True
+        local_background : bool, optional
+            Estimate and remove a local background, by default True
         background_clipping : int, optional
             Sigma clipping used for the background, by default 3
         grouping_threshold : int, optional
@@ -933,6 +936,7 @@ class ContMap(NDDataArray):
             sources["eflux_peak"] = Column(self.uncertainty.array[y_idx, x_idx], unit=self.unit * u.beam)
 
         if psf:
+            data = self.data
             # BasicPSFPhotometry with fixed positions
 
             sigma_psf = self.beam.stddev_maj.to(u.pix, self._pixel_scale).value
@@ -951,9 +955,13 @@ class ContMap(NDDataArray):
             if len(xx) > 1:
                 source_grouper = SourceGrouper(grouping_threshold * self.beam.major.to(u.pix, self._pixel_scale).value)
 
-            local_bkg = None
+            bkgstat = MedianBackground(sigma_clip=SigmaClip(sigma=background_clipping, stdfunc="mad_std"))
+
             if background:
-                bkgstat = MedianBackground(sigma_clip=SigmaClip(sigma=background_clipping, stdfunc="mad_std"))
+                data = data - bkgstat(data)
+
+            local_bkg = None
+            if local_background:
                 local_bkg = LocalBackground(5, 10, bkgstat)
 
             photometry = PSFPhotometry(
@@ -969,7 +977,7 @@ class ContMap(NDDataArray):
             )
 
             result_tab = photometry(
-                self.data,
+                data,
                 mask=self.mask,
                 error=1 / np.sqrt(self.weights),
                 init_params=positions,
@@ -990,7 +998,7 @@ class ContMap(NDDataArray):
                 # Transform pixel coordinates column to world coordinates
                 sources = xy_to_world(sources, self.wcs, "x_fit", "y_fit")
 
-            self._residual = photometry.make_residual_image(self.data, (10, 10))
+            self._residual = photometry.make_residual_image(data, (10, 10))
 
         self.sources = sources